Add inventory system
Реализовано: 1. Инвентарь 2. Используемые расходники. В доработке: Экипировка, интерфейс, смена экипировки.

+++ Test idle/Assets/Scriptable object/Monsters/MonstersAbstract.cs
@@ -3,7 +3,7 @@ public class MonstersAbstract : ScriptableObject
 {
     [SerializeField] protected string nameMonster;
     public string _nameMonster => nameMonster;
-    public int hitPoint;
-    public int damage;
+    public float hitPoint;
+    public float damage;
     public float attackSpeed;
 }
