+++ Test idle/Assets/Scriptable object/Player/PlayerSettings.cs
@@ -3,29 +3,29 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Player", menuName = "PlayerCreate", order = 0)]
 public class PlayerSettings : ScriptableObject
 {
-    [SerializeField] private int hitPoint;
-    public int _hitPoint => hitPoint;
+    [SerializeField] private float hitPoint;
+    public float _hitPoint => hitPoint;
     
-    [SerializeField] private int damage;
-    public int _damage => damage;
+    [SerializeField] private float damage;
+    public float _damage => damage;
 
-    [SerializeField] private int armour;
-    public int _armour => armour;
+    [SerializeField] private float armour;
+    public float _armour => armour;
 
     [SerializeField] private float attackSpeed;
     public float _attackSpeed => attackSpeed;
 
-    public void UpgradeHitPoint(int buffHp)
+    public void UpgradeHitPoint(float buffHp)
     {
         hitPoint += buffHp;
     }
 
-    public void UpgradeDamage(int buffDamage)
+    public void UpgradeDamage(float buffDamage)
     {
         damage += buffDamage;
     }
 
-    public void UpgradeArmour(int buffArmour)
+    public void UpgradeArmour(float buffArmour)
     {
         armour += buffArmour;
     }
